+++ frontend/components/Chain.vue
@@ -172,18 +172,6 @@
           </div>
         </div>
       </div>
-      <div class="col-6 col-md-6 col-lg-3 mb-4">
-        <div class="card h-100">
-          <div class="card-body">
-            <h4 class="mb-3 title">
-              {{ $t('components.network.transactionsFees') }}
-            </h4>
-            <h6 class="d-inline-block">
-              {{ formatAmount(totalTransactionsFees) }}
-            </h6>
-          </div>
-        </div>
-      </div>
     </div>
   </div>
 </template>
@@ -272,9 +260,6 @@ export default {
           }
         `,
         result({ data }) {
-          this.totalTransactionsFees =
-            data.total.find((row) => row.name === 'transactions_fees').count ||
-            0
           this.totalExtrinsics =
             data.total.find((row) => row.name === 'extrinsics').count || 0
           this.totalSignedExtrinsics =